Number => Real when declaring isless
DualNumbers.jl (used by Gadfly) and SIUnits declare ambiguous isless
function. Since isless is only meaningfull for real numbers, and since
both DualNumbers and SIUnit declare it for Number rather than Real
arguments, the trick is for SIUnit to downgrade its declaration to the
more specific type. Number is a supertype of both Real and Complex.